changes in comment popup
